% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genetic_data.R
\name{get_genetic_data}
\alias{get_genetic_data}
\title{Generate Genetic Data}
\usage{
get_genetic_data(
  n,
  prob = 0.5,
  n_unif,
  min,
  max,
  prob_X = 0.5,
  n_norm,
  mean = 0,
  sd = 0.5
)
}
\arguments{
\item{n}{Number of samples to generate data for.}

\item{prob}{Probability for generating binary confounders.}

\item{n_unif}{Number of uniform random variables to generate for minor allele frequency (MAF).}

\item{min}{Minimum value for the uniform distribution used in generating MAF.}

\item{max}{Maximum value for the uniform distribution used in generating MAF.}

\item{prob_X}{Probability for the effect of confounders on the outcome.}

\item{n_norm}{Number of normal random variables to generate for the genetic effect.}

\item{mean}{Mean value for the normal distribution used in generating the genetic effect.}

\item{sd}{Sd for the normal distribution used in generating the genetic effect.}
}
\value{
A data frame containing three variables:
\itemize{
\item Y: binary outcome variable. a vector of length n.
\item X: d confounders, either a nx1 vector or nxd matrix if d>1.
\item G: genotypes of varaints. a n*p matrix.
}
}
\description{
This function generates genetic data for a given number of samples.
It simulates binary outcome variables, confounders, and genotypes of variants based on the specified parameters.
}
\examples{
# To generate genetic data for 2000 samples:

set.seed(1234)
result <- get_genetic_data(n = 2000, n_unif=10, min=0.001,max=0.01, n_norm=10)
head(result)
}
