% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Summarize MiST Analysis Results in a Table}
\usage{
get_table(res)
}
\arguments{
\item{res}{A list for the result of a MiST analysis, expected to include 'stat' and 'pvalue'.}
}
\value{
A kable of MiST analysis results.
}
\description{
This function takes the result of a MiST analysis and summarizes the estimates and pvalues in a table format.
}
\examples{
# Data generation
n = 2000
set.seed(1234)
X = rbinom(n,size=1,prob=0.5)
MAF = runif(10,min=0.001,max=0.01)
G = sapply(MAF,function(maf) rbinom(n,size=1,prob=maf))
eta = X*0.5 + G\%*\%rnorm(10,mean=0,sd=0.5)
Y = rbinom(n,size=1,prob=exp(eta)/(1+exp(eta)))
d = 1
p = 10

data = data.frame(Y=Y, X=X,G=G)
result <- MiST(data = data,
         outcome_type = "Binary",
         d = d,
         p = p,
         R = 1,
         weight_method = "User")

get_table(result)
}
