% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiST.R
\name{MiST}
\alias{MiST}
\title{Implementation of set-based mixed effects score tests via \emph{MiST}.}
\usage{
MiST(
  data,
  d = 0,
  p,
  R = 1,
  outcome_type = "Continuous",
  weight_method = "User",
  user_weight = rep(1, p),
  burden_scale = TRUE,
  chisq_app = "3M",
  combinations_return = TRUE,
  combination_preference = "All",
  acc = 5e-10,
  acc_auto = TRUE,
  accurate_app_threshold = -log10(0.05),
  max_core = 4
)
}
\arguments{
\item{data}{A dataframe \code{[Y,X,G]} with rows represent subjects and columns are \code{Y} (outcome), \code{X} (confounders), and \code{G} (genotypes).}

\item{d}{Number of confounders. If \code{d} is 0, there is no confounder, and the input data should be arranged as \code{[Y G]}.}

\item{p}{Number of genetic variants in the set.}

\item{R}{Number of burden scores (R>=0). When \code{R} is 0, only the variance component test is performed and more than one variants are required in this situation.}

\item{outcome_type}{Either "Continuous" for quantitative trait, or "Binary" for dichotomous trait.}

\item{weight_method}{Specifying weights for calculating weighted burden scores. It takes "User" or "No weight". "User"(default) allows user-defined weights for each variant (e.g., functional annotations and equal weights), which are specified by the input option "user_weight". "No weight" sets all weights = 0 and tests only the variance component.}

\item{user_weight}{A vector or a matrix specifying weights for calculating weighted burden scores. This option only works when the weight_method is set as "User". If the weight_method is set as "User"" and no user_weight is specified, the weight is set as 1.}

\item{burden_scale}{A logical value (default: TRUE) specifying if standardization on burden components needs to be carried out when there are more than 1 burden score.}

\item{chisq_app}{Either "3M" (default) or "4M" for the moment matching (Liu's) method in the quantile approximation in the optimal linear combination of burden and variance components. "3M" matches the 3rd moment and "4M" matches the 4th moment of the target and approximate distributions.}

\item{combinations_return}{A logical value with TRUE (default) indicating that the tests for combining the burden and variance components are performed. The combined methods include the optimal linear combination, data-adpative weighted combination, and Fisher's combination. If it is FALSE, only the burden and variance component tests are returned.}

\item{combination_preference}{Either of "All" (default), or a vector containing "OptMin", "AdptWt", or "Fisher" to specify the combination method(s).}

\item{acc}{A numerical value indicating the precision of the Davies method for p-value calculation. Default is 5e-10.}

\item{acc_auto}{A logical value (default: TRUE) indicating if data adaptive precision is used in optimal linear combination. We recommend to set this as TRUE for computational efficiency.}

\item{accurate_app_threshold}{A numerical value specifying the threshold to determine when the Liu or Davies method is used in the quantile approximation in the optimal linear combination of burden and variance components. Default is -log10(0.05).}

\item{max_core}{An integer specifying the maximum number of cores that can be recruited in the parallel package. Default is 4 cores.}
}
\value{
A list containing \code{pvalue}, \code{stat}, \code{pvalue.ind.f}, \code{stat.ind.f}, \code{rho}, and \code{data.info}.
Specifically, \code{pvalue} are obtained from the burden and variance components tests, as well as three combination tests including optimal linear combination \emph{oMiST}, data-adaptive weighted combination \emph{aMiST}, and Fiser's combination \emph{fMiST}.
\code{stat} are the corresponding test statistics.
\code{pvalue.ind.f} and \code{stat.ind.f} are p-values and test statistics of individual burden scores if multiple weights are included.
\code{rho} is the optimal weight of the burden and random-effects components in \emph{oMiST} and \emph{aMiST}.
\code{data.info} includes the numbers of genetic variants and burden scores.
}
\description{
Implementation of set-based mixed effects score tests via \emph{MiST}.
}
\examples{
# Data generation
n = 2000
set.seed(1234)
X = rbinom(n,size=1,prob=0.5)
MAF = runif(10,min=0.001,max=0.01)
G = sapply(MAF,function(maf) rbinom(n,size=1,prob=maf))
eta = X*0.5 + G\%*\%rnorm(10,mean=0,sd=0.5)
Y = rbinom(n,size=1,prob=exp(eta)/(1+exp(eta)))
d = 1
p = 10

data = data.frame(Y=Y, X=X,G=G)
mist = MiST(data = data,
         outcome_type = "Binary",
         d = d,
         p = p,
         R = 1,
         weight_method = "User")
mist
}
