% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_approximation.R
\name{pvalue.liu}
\alias{pvalue.liu}
\title{Calculate p-Value for weighted sum of chisq Distributions}
\usage{
pvalue.liu(Q, weight, method = "3M", df = rep(1, length(weight)))
}
\arguments{
\item{Q}{A test statistic for which the p-value is to be calculated.}

\item{weight}{A vector of weights for each chisq}

\item{method}{The method used for approximation,
default: "3M". Other options include: "4M" which is a modified version of the \code{WeightSumChisq} function.}

\item{df}{A vector of df for each chisq, default: 1 for each weight}
}
\value{
A list containing the \code{pvalue}, the approximated \code{df}, and the noncentrality parameter \code{ncp}
}
\description{
This function calculates the p-value for a given test statistic \code{Q} that is a weighted sum of chisq distributions.
It uses Liu's(2009) approximation method to adjust \code{Q} and calculates the p-value based on the noncentral chisq.
}
\examples{
# Define the statistic Q, weights, and degrees of freedom for the chi-squared distributions
Q = 10.5
weights = c(0.5, 1.5, 2.0)
dfs = c(1, 1, 1) # Assuming all chi-squared distributions have 1 degree of freedom

result = pvalue.liu(Q, weights, df = dfs)
result
}
