% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_approximation.R
\name{WeightSumChisq}
\alias{WeightSumChisq}
\title{Approximate Noncentral chisq distribution for the weighted sum of independent chisq distributions}
\usage{
WeightSumChisq(weight, df = rep(1, length(weight)))
}
\arguments{
\item{weight}{A vector of weights for each chisq distribution.}

\item{df}{A vector of df for each chisq distribution, default: 1 for each weight}
}
\value{
A vector of mean, sd of the weighted sum, the df(\code{l}) and
the noncentrality parameter of the approximation noncentral chisq
}
\description{
This function calculates the df and the noncentral parameter of an approximated noncentral chisq distribution for the weighted sum of independent chisq distributions.
}
\examples{
# Define weights for the chi-squared distributions
weights <- c(0.5, 1.5, 2.0)

result <- WeightSumChisq(weight = weights)
result
}
